+++ R/PDE.R
@@ -27,10 +27,11 @@
 #'
 #' @docType package
 #' @name PDE
+#' @aliases PDE-package
 NULL
 #> NULL
 
-## 1.4.6
+## 1.4.7
 
 ## TODO save progress with tsv file
 ## TODO export as excel file
@@ -1218,8 +1219,27 @@ PDE_install_Xpdftools4.02 <- function(sysname=NULL, bin=NULL, verbose=TRUE, perm
     output.without.html <- gsub("&divide;",intToUtf8(0x00F7),output.without.html)
     output.without.html <- gsub("&circ;",intToUtf8(0x00FE),output.without.html)
     output.without.html <- gsub("&tilde;",intToUtf8(0x007E),output.without.html)
-    output.without.html <- gsub(intToUtf8(0xFB01),"fi",output.without.html, fixed = TRUE)
     return(output.without.html)
+  }
+  
+  replace.letter.merge <- function(input.with.symbol, test = TRUE) {
+    output.without.symbol <- input.with.symbol
+    if (test == TRUE){
+      if (!any(grepl(paste0(intToUtf8(0xFB01),"|",intToUtf8(0xFB02),"|",
+                           intToUtf8(0xFB03),"|",intToUtf8(0xFB04),"|",
+                           intToUtf8(0xFB05),"|",intToUtf8(0xFB06)), input.with.symbol))){
+        test <- FALSE
+      } 
+    }
+    if (test == TRUE){
+      output.without.symbol <- gsub(intToUtf8(0xFB01),"fi",output.without.symbol, fixed = TRUE)
+      output.without.symbol <- gsub(intToUtf8(0xFB02),"fl",output.without.symbol, fixed = TRUE)
+      output.without.symbol <- gsub(intToUtf8(0xFB03),"ffi",output.without.symbol, fixed = TRUE)
+      output.without.symbol <- gsub(intToUtf8(0xFB04),"ffl",output.without.symbol, fixed = TRUE)
+      output.without.symbol <- gsub(intToUtf8(0xFB05),"ft",output.without.symbol, fixed = TRUE)
+      output.without.symbol <- gsub(intToUtf8(0xFB06),"st",output.without.symbol, fixed = TRUE)
+    }
+    return(output.without.symbol)
   }
 
   insert.html.entity <- function(input.without.html) {
@@ -2036,7 +2056,7 @@ PDE_install_Xpdftools4.02 <- function(sysname=NULL, bin=NULL, verbose=TRUE, perm
         if (grepl("&",line.txthtmlcontent)){
           txthtmlcontent[[j]][z] <- replace.html.entity(line.txthtmlcontent)
         } else {
-          txthtmlcontent[[j]][z] <- line.txthtmlcontent
+          txthtmlcontent[[j]][z] <- replace.letter.merge(line.txthtmlcontent, test = TRUE)
         }
       }  ## end go through each line z
       content[[(j+2)]] <- txthtmlcontent[[j]]
@@ -2305,6 +2325,8 @@ PDE_install_Xpdftools4.02 <- function(sysname=NULL, bin=NULL, verbose=TRUE, perm
       ignore.case.sw <- search.word.table[i, "ignore.case.sw"]
       word.txtline <- NULL
       word.keeplayoutline <- NULL
+      ##TODO
+      ## Use result <- str_detect(df$column_name, "pattern")?
       ## if search words were not chosen write all lines in txtline
       if (search.word.table[i, "words"] == ""){
         word.txtline <- 1:length(txtcontent)
@@ -3769,6 +3791,7 @@ PDE_install_Xpdftools4.02 <- function(sysname=NULL, bin=NULL, verbose=TRUE, perm
     word.table <- NULL
     ## start search only when search words are in the
     ## list
+    processed.tab.category <- NULL
     if (!is.na(search.word.table[1, "words"])) {
       if (search.word.table[1, "words"] != "") {
         pdf_searchwords <- NULL
@@ -9662,16 +9685,16 @@ PDE_reader_i <- function(verbose=TRUE) {
 
     ## line2
     load.pdffolder <- function(pdffolder) {
-      catchvariable <<- tcltk::tclvalue(pdffolder.var)
-      catchvariable <<- pdffolder
+      catchvariable <- tcltk::tclvalue(pdffolder.var)
+      catchvariable <- pdffolder
       if (length(pdffolder) == 0 || pdffolder == "%pdffolder") {
         tcltk::tclvalue(pdffolder.var) <- tcltk::tk_choose.dir(default = tcltk::tclvalue(pdffolder.var),
                                                                caption = "Choose folder with PDF files")
-        catchvariable <<- tcltk::tclvalue(pdffolder.var)
+        catchvariable <- tcltk::tclvalue(pdffolder.var)
       }
       ## search pdf folder for matching PDF files
       if (!tcltk::tclvalue(jumpto.var) == "") {
-        catchvariable <<- tcltk::tclvalue(jumpto.var)
+        catchvariable <- tcltk::tclvalue(jumpto.var)
         analysis.filename <- gsub("^[^-]*-", "", tcltk::tclvalue(jumpto.var))
         if (substr(analysis.filename,1,2) == "!_" ||
             substr(analysis.filename,1,2) == "x_"){
@@ -9693,11 +9716,11 @@ PDE_reader_i <- function(verbose=TRUE) {
                             ".pdf")
         }
         ## find pdf
-        catchvariable <<- pdfname
+        catchvariable <- pdfname
         pdfs <- list.files(tcltk::tclvalue(pdffolder.var),
                            pattern = "*.pdf", full.names = TRUE,
                            recursive = TRUE)
-        catchvariable <<- any(grepl(pdfname, pdfs, fixed = TRUE))
+        catchvariable <- any(grepl(pdfname, pdfs, fixed = TRUE))
         if (any(grepl(pdfname, pdfs, fixed = TRUE))) {
           tcltk::tclvalue(pdfname.var) <- basename(grep(pdfname,
                                                         pdfs, fixed = TRUE, value = TRUE)[1])